Force daemon.R to preload worker functions before forking them.
A worker may be called many times. It turns out that much of the time of
a worker is spent loading libraries.  The first time daemon.R sees a
worker it will run it inline. In other words, it will not fork worker
functions that it hasn't seen before but rather runs the worker
directly. After the first worker finishes, all remaining calls will be
forked, as before. This way we won't keep reloading library
dependencies.

The other two changes here are

1) we increased the timeout to the daemon from 10 seconds to 5 minutes.
This means if a worker function takes longer than 5 minutes, the daemon
will fail.

2) We load worker.R only once as it now defines a function,
MaybeForkedWorker.